refactor: extract app context initialization
